feat: use server-side apply
