+++ atmosphere/flows.py
@@ -1,7 +1,8 @@
 from taskflow.patterns import graph_flow
 
 from atmosphere.config import CONF
-from atmosphere.tasks import flux, kubernetes, openstack_helm
+from atmosphere.tasks.composite import openstack_helm
+from atmosphere.tasks.kubernetes import flux, v1
 
 NAMESPACE_CERT_MANAGER = "cert-manager"
 NAMESPACE_KUBE_SYSTEM = "kube-system"
@@ -27,36 +28,40 @@ NODE_FEATURE_DISCOVERY_VALUES = {
     "master": {"nodeSelector": CONTROL_PLANE_NODE_SELECTOR}
 }
 
+PERCONA_XTRADB_OPERATOR_VALUES = {
+    "nodeSelector": CONTROL_PLANE_NODE_SELECTOR,
+}
+
 
 def get_deployment_flow():
     flow = graph_flow.Flow("deploy").add(
         # kube-system
-        kubernetes.CreateOrUpdateNamespaceTask(name=NAMESPACE_KUBE_SYSTEM),
-        flux.CreateOrUpdateHelmRepositoryTask(
+        v1.ApplyNamespaceTask(name=NAMESPACE_KUBE_SYSTEM),
+        flux.ApplyHelmRepositoryTask(
             namespace=NAMESPACE_KUBE_SYSTEM,
             name=HELM_REPOSITORY_CEPH,
             url="https://ceph.github.io/csi-charts",
         ),
         # cert-manager
-        kubernetes.CreateOrUpdateNamespaceTask(name=NAMESPACE_CERT_MANAGER),
-        flux.CreateOrUpdateHelmRepositoryTask(
+        v1.ApplyNamespaceTask(name=NAMESPACE_CERT_MANAGER),
+        flux.ApplyHelmRepositoryTask(
             namespace=NAMESPACE_CERT_MANAGER,
             name=HELM_REPOSITORY_JETSTACK,
             url="https://charts.jetstack.io",
         ),
         # monitoring
-        kubernetes.CreateOrUpdateNamespaceTask(name=NAMESPACE_MONITORING),
-        flux.CreateOrUpdateHelmRepositoryTask(
+        v1.ApplyNamespaceTask(name=NAMESPACE_MONITORING),
+        flux.ApplyHelmRepositoryTask(
             namespace=NAMESPACE_MONITORING,
             name=HELM_REPOSITORY_PROMETHEUS_COMMUINTY,
             url="https://prometheus-community.github.io/helm-charts",
         ),
-        flux.CreateOrUpdateHelmRepositoryTask(
+        flux.ApplyHelmRepositoryTask(
             namespace=NAMESPACE_MONITORING,
             name=HELM_REPOSITORY_NODE_FEATURE_DISCOVERY,
             url="https://kubernetes-sigs.github.io/node-feature-discovery/charts",
         ),
-        flux.CreateOrUpdateHelmReleaseTask(
+        flux.ApplyHelmReleaseTask(
             namespace=NAMESPACE_MONITORING,
             name="node-feature-discovery",
             repository=HELM_REPOSITORY_NODE_FEATURE_DISCOVERY,
@@ -65,33 +70,41 @@ def get_deployment_flow():
             values=NODE_FEATURE_DISCOVERY_VALUES,
         ),
         # openstack
-        kubernetes.CreateOrUpdateNamespaceTask(name=NAMESPACE_OPENSTACK),
-        flux.CreateOrUpdateHelmRepositoryTask(
+        v1.ApplyNamespaceTask(name=NAMESPACE_OPENSTACK),
+        flux.ApplyHelmRepositoryTask(
             namespace=NAMESPACE_OPENSTACK,
             name=HELM_REPOSITORY_BITNAMI,
             url="https://charts.bitnami.com/bitnami",
         ),
-        flux.CreateOrUpdateHelmRepositoryTask(
+        flux.ApplyHelmRepositoryTask(
             namespace=NAMESPACE_OPENSTACK,
             name=HELM_REPOSITORY_PERCONA,
             url="https://percona.github.io/percona-helm-charts/",
         ),
-        flux.CreateOrUpdateHelmRepositoryTask(
+        flux.ApplyHelmReleaseTask(
+            namespace=NAMESPACE_OPENSTACK,
+            name="pxc-operator",
+            repository=HELM_REPOSITORY_PERCONA,
+            chart="pxc-operator",
+            version="1.10.0",
+            values=PERCONA_XTRADB_OPERATOR_VALUES,
+        ),
+        flux.ApplyHelmRepositoryTask(
             namespace=NAMESPACE_OPENSTACK,
             name=HELM_REPOSITORY_INGRESS_NGINX,
             url="https://kubernetes.github.io/ingress-nginx",
         ),
-        flux.CreateOrUpdateHelmRepositoryTask(
+        flux.ApplyHelmRepositoryTask(
             namespace=NAMESPACE_OPENSTACK,
             name=HELM_REPOSITORY_OPENSTACK_HELM_INFRA,
             url="https://tarballs.opendev.org/openstack/openstack-helm-infra/",
         ),
-        flux.CreateOrUpdateHelmRepositoryTask(
+        flux.ApplyHelmRepositoryTask(
             namespace=NAMESPACE_OPENSTACK,
             name=HELM_REPOSITORY_COREDNS,
             url="https://coredns.github.io/helm",
         ),
-        flux.CreateOrUpdateHelmRepositoryTask(
+        flux.ApplyHelmRepositoryTask(
             namespace=NAMESPACE_OPENSTACK,
             name=HELM_REPOSITORY_OPENSTACK_HELM,
             url="https://tarballs.opendev.org/openstack/openstack-helm/",
@@ -100,16 +113,16 @@ def get_deployment_flow():
 
     if CONF.memcached.enabled:
         flow.add(
-            openstack_helm.CreateOrUpdateReleaseSecretTask(
+            openstack_helm.ApplyReleaseSecretTask(
                 namespace=NAMESPACE_OPENSTACK, chart="memcached"
             ),
-            openstack_helm.CreateOrUpdateHelmReleaseTask(
+            openstack_helm.ApplyHelmReleaseTask(
                 namespace=NAMESPACE_OPENSTACK,
                 repository=HELM_REPOSITORY_OPENSTACK_HELM_INFRA,
                 name="memcached",
                 version="0.1.12",
             ),
-            kubernetes.CreateOrUpdateServiceTask(
+            v1.ApplyServiceTask(
                 namespace=NAMESPACE_OPENSTACK,
                 name="memcached-metrics",
                 labels={
