+++ atmosphere/tasks/composite/openstack_helm.py
@@ -1,8 +1,8 @@
 from atmosphere.models.openstack_helm import values
-from atmosphere.tasks import flux, kubernetes
+from atmosphere.tasks.kubernetes import flux, v1
 
 
-class CreateOrUpdateReleaseSecretTask(kubernetes.CreateOrUpdateSecretTask):
+class ApplyReleaseSecretTask(v1.ApplySecretTask):
     def __init__(self, namespace: str, chart: str, *args, **kwargs):
         super().__init__(
             namespace,
@@ -13,7 +13,7 @@ class CreateOrUpdateReleaseSecretTask(kubernetes.CreateOrUpdateSecretTask):
         )
 
 
-class CreateOrUpdateHelmReleaseTask(flux.CreateOrUpdateHelmReleaseTask):
+class ApplyHelmReleaseTask(flux.ApplyHelmReleaseTask):
     def __init__(
         self,
         namespace: str,
